requirements.txt
Django==2.2
uWSGI==2.0.19 # used to run python applications on production
# gunicorn>=20.0.4,<20.1
SQLAlchemy==1.3.13
numpy==1.17.2
pandas==0.25.1

plotly==4.2.0
dash==1.11.0
dash-bootstrap-components==0.11.3
dash-core-components==1.9.1
dash-daq==0.3.3
dash-html-components==1.0.3
dash-renderer==1.4.0
dash-table==4.6.2
cryptography

PyMySQL==0.10.0
xlrd==2.0.1
Flask-Caching==1.8.0
xlsxwriter
regex
psycopg2>=2.8.6,<2.9

# ===========================================

# Django==3.2.4
# uWSGI>=2.0.19.1,<2.1 # used to run python applications on production
# # gunicorn>=20.0.4,<20.1
# SQLAlchemy==1.4.18
# numpy==1.20.3
# pandas==1.2.4

# plotly==4.14.3
# dash==1.20.0
# dash-bootstrap-components==0.12.2
# dash-core-components==1.16.0
# dash-daq==0.5.0
# dash-html-components==1.1.3
# dash-renderer==1.9.1
# dash-table==4.11.3
# cryptography

# PyMySQL==1.0.2
# xlrd==2.0.1
# Flask-Caching==1.8.0
# xlsxwriter
# regex
# psycopg2>=2.8.6,<2.9